move project action to dedicated route file
